Make SG config UI use generic i18n modules.
