+++ public/apps/configuration-react/components/Icon/assets/authc_and_authz_dark.js
@@ -3,7 +3,7 @@ import React from 'react';
 const SvgAuthcAndAuthzDark = props => (
   <svg id="authc_and_authz_dark_svg__ART" viewBox="0 0 48 48" {...props}>
     <defs>
-      <style>{".authc_and_authz_dark_svg__cls-2{fill:#047268}"}</style>
+      <style>{'.authc_and_authz_dark_svg__cls-2{fill:#047268}'}</style>
     </defs>
     <path
       d="M33.44 21.45H14.56a3.32 3.32 0 0 0-3.32 3.31v13.79a3.32 3.32 0 0 0 3.32 3.31h18.88a3.32 3.32 0 0 0 3.32-3.31V24.76a3.32 3.32 0 0 0-3.32-3.31zm-7.29 10.84a1.5 1.5 0 0 0-.31.94v.41a1.84 1.84 0 1 1-3.68 0v-.41a1.48 1.48 0 0 0-.3-.93 2.75 2.75 0 1 1 4.9-1.72 2.7 2.7 0 0 1-.61 1.71z"
