+++ public/apps/configuration-react/components/Icon/assets/logo_dark.js
@@ -3,7 +3,7 @@ import React from 'react';
 const SvgLogoDark = props => (
   <svg id="logo_dark_svg__ART" viewBox="0 0 48 48" {...props}>
     <defs>
-      <style>{".logo_dark_svg__cls-1{fill:#fff}"}</style>
+      <style>{'.logo_dark_svg__cls-1{fill:#fff}'}</style>
     </defs>
     <path
       className="logo_dark_svg__cls-1"
