+++ public/apps/configuration-react/components/Icon/assets/purge_cache.js
@@ -3,7 +3,7 @@ import React from 'react';
 const SvgPurgeCache = props => (
   <svg id="purge_cache_svg__ART" viewBox="0 0 48 48" {...props}>
     <defs>
-      <style>{".purge_cache_svg__cls-2{fill:#353740}"}</style>
+      <style>{'.purge_cache_svg__cls-2{fill:#353740}'}</style>
     </defs>
     <circle
       cx={24}
