+++ public/apps/configuration-react/components/Icon/assets/purge_cache_dark.js
@@ -3,7 +3,7 @@ import React from 'react';
 const SvgPurgeCacheDark = props => (
   <svg id="purge_cache_dark_svg__ART" viewBox="0 0 48 48" {...props}>
     <defs>
-      <style>{".purge_cache_dark_svg__cls-2{fill:#fff}"}</style>
+      <style>{'.purge_cache_dark_svg__cls-2{fill:#fff}'}</style>
     </defs>
     <circle
       cx={24}
