+++ public/apps/configuration-react/components/Icon/assets/tenant_pattern.js
@@ -5,7 +5,7 @@ const SvgTenantPattern = props => (
     <defs>
       <style>
         {
-          ".tenant_pattern_svg__cls-1{fill:#047268}.tenant_pattern_svg__cls-2{fill:#353740}"
+          '.tenant_pattern_svg__cls-1{fill:#047268}.tenant_pattern_svg__cls-2{fill:#353740}'
         }
       </style>
     </defs>
