+++ public/apps/configuration-react/components/Icon/assets/tenants.js
@@ -3,7 +3,7 @@ import React from 'react';
 const SvgTenants = props => (
   <svg id="tenants_svg__ART" viewBox="0 0 48 48" {...props}>
     <defs>
-      <style>{".tenants_svg__cls-1{fill:#353740}"}</style>
+      <style>{'.tenants_svg__cls-1{fill:#353740}'}</style>
     </defs>
     <circle className="tenants_svg__cls-1" cx={12} cy={12.75} r={3.75} />
     <circle className="tenants_svg__cls-1" cx={24} cy={12.75} r={3.75} />
