+++ public/apps/configuration-react/components/Icon/assets/tenants_dark.js
@@ -3,7 +3,7 @@ import React from 'react';
 const SvgTenantsDark = props => (
   <svg id="tenants_dark_svg__ART" viewBox="0 0 48 48" {...props}>
     <defs>
-      <style>{".tenants_dark_svg__cls-1{fill:#fff}"}</style>
+      <style>{'.tenants_dark_svg__cls-1{fill:#fff}'}</style>
     </defs>
     <circle className="tenants_dark_svg__cls-1" cx={12} cy={12.75} r={3.75} />
     <circle className="tenants_dark_svg__cls-1" cx={24} cy={12.75} r={3.75} />
