+++ public/apps/configuration-react/utils/i18n/common.js
@@ -1,49 +1,1 @@
-import React from 'react';
-import { EuiI18n } from '@elastic/eui';
-
-export const typeText = (<EuiI18n token="sg.common.type.text" default="Type" />);
-export const descriptionText = (<EuiI18n token="sg.common.description.text" default="Description" />);
-export const jsonText = (<EuiI18n token="sg.common.json.text" default="JSON" />);
-export const saveText = (<EuiI18n token="sg.common.save.text" default="Save" />);
-export const cancelText = (<EuiI18n token="sg.common.cancel.text" default="Cancel" />);
-export const confirmText = (<EuiI18n token="sg.common.confirm.text" default="Confirm" />);
-export const addText = (<EuiI18n token="sg.common.add.text" default="Add" />);
-export const removeText = (<EuiI18n token="sg.common.remove.text" default="Remove" />);
-export const noItemsFoundText = (<EuiI18n token="sg.common.noItemsFound.text" default="No items found" />);
-export const requiredText = (<EuiI18n token="sg.common.required.text" default="Required" />);
-export const errorText = (<EuiI18n token="sg.common.error.text" default="Error" />);
-export const inspectText = (<EuiI18n token="sg.common.inspect.text" default="Inspect" />);
-export const closeText = (<EuiI18n token="sg.common.close.text" default="Close" />);
-export const nameText = (<EuiI18n token="sg.common.name.text" default="Name" />);
-export const reservedText = (<EuiI18n token="sg.common.reserved.text" default="Reserved" />);
-export const cloneText = (<EuiI18n token="sg.common.clone.text" default="Clone" />);
-export const deleteText = (<EuiI18n token="sg.common.delete.text" default="Delete" />);
-export const confirmDeleteText = (<EuiI18n token="sg.common.confirmDelete.text" default="Confirm Delete" />);
-export const doYouReallyWantToDeleteText = (
-  <EuiI18n token="sg.common.doYouReallyWantToDelete.text" default="Do you really want to delete" />
-);
-export const currentUserText = (<EuiI18n token="sg.common.currentUser.text" default="Current User" />);
-export const navigateText = (<EuiI18n token="sg.common.navigate.text" default="Navigate" />);
-export const disabledText = (<EuiI18n token="sg.common.disabled.text" default="Disabled" />);
-export const successText = (<EuiI18n token="sg.common.success.text" default="Success" />);
-export const problemWithValidationTryAgainText = (
-  <EuiI18n token="sg.common.problemWithValidationTryAgain.text" default="Problem with validation. Please try again" />
-);
-export const nameAlreadyExistsText = (
-  <EuiI18n token="sg.tenants.nameAlreadyExists.text" default="Name already exists" />
-);
-export const nameMustNotContainDotsText = (
-  <EuiI18n token="sg.internalUsers.nameMustNotContainDots.text" default="Name must not contain '.'" />
-);
-export const permissionsText = (<EuiI18n token="sg.common.permissions.text" default="Permissions" />);
-export const theOptionIsNotValidText = (<EuiI18n token="sg.common.theOptionIsNotValid.text" default="The option is not valid" />);
-export const advancedText = (<EuiI18n token="sg.common.advanced.text" default="Advanced" />);
-export const includeText = (<EuiI18n token="sg.common.include.text" default="Include" />);
-export const excludeText = (<EuiI18n token="sg.common.exclude.text" default="Exclude" />);
-export const jsonIsInvalidText = (<EuiI18n token="sg.common.jsonIsInvalid.text" default="JSON is invalid" />);
-export const jsonMustNotBeEmptyText = (<EuiI18n token="sg.common.jsonMustNotBeEmpty.text" default="JSON must not be empty" />);
-export const occupiedText = (<EuiI18n token="sg.common.occupied.text" default="Occupied" />);
-export const availableText = (<EuiI18n token="sg.common.available.text" default="Available" />);
-export const systemItemsText = (<EuiI18n token="sg.common.systemItems.text" default="System Items" />);
-export const logoutText = (<EuiI18n token="sg.common.logout.text" default="Logout" />);
-export const loginText = (<EuiI18n token="sg.common.login.text" default="Login" />);
+export * from '../../../utils/i18n/common';
